Working version of sorts.
Still requires search implemented properly, and this may need wrapped
into a form so that we're not messing about regarding form submission.
DB conn is also very flaky, but not sure if this is localhost setup or
what. Seems intermittent after long periods of inactivity. Not sure if
some pooling is timing out or what.